LayoutTests/imported/w3c/web-platform-tests/dom/events/EventListener-handleEvent-cross-realm-expected.txt



PASS EventListener is cross-realm plain object without 'handleEvent' property
PASS EventListener is cross-realm plain object with non-callable 'handleEvent' property
FAIL EventListener is cross-realm plain object with revoked Proxy as 'handleEvent' property assert_equals: expected function "function TypeError() {
    [native code]
}" but got function "function TypeError() {
    [native code]
}"
PASS EventListener is cross-realm non-callable revoked Proxy
PASS EventListener is cross-realm callable revoked Proxy
